+++ src/main/java/com/example/service/DepartmentUpdateService.java
@@ -3,6 +3,6 @@ package com.example.service;
 import com.example.entity.Department;
 
 public interface DepartmentUpdateService {
-	Department updateData(Department department , int id);
+	Department updateData(Department department , Long id);
 
 }
